Move elf loading to different file
